Finish(?) can-bind’s initial implementation
- Prevent listeners of bindings that have been turned off from updating values (similar to https://github.com/canjs/can-stache-bindings/pull/459)
- Change how one-way vs. two-way binding detection occurs (now, always accept the option for what it is, otherwise check different symbols on both values)
- Fix naming the updateChild and updateParent functions
- Remove debug dependencies in production
- Add a warning when an update is being ignored because of the cycles detection
- Add more docs & tests
