+++ can-bind.js
@@ -1,10 +1,15 @@
 var canReflect = require("can-reflect");
-var canReflectDeps = require("can-reflect-dependencies");
 var canSymbol = require("can-symbol");
 var queues = require("can-queues");
 
+//!steal-remove-start
+var canLog = require("can-log/dev/dev");
+var canReflectDeps = require("can-reflect-dependencies");
+//!steal-remove-end
+
 // Symbols
 var getChangesSymbol = canSymbol.for("can.getChangesDependencyRecord");
+var getValueSymbol = canSymbol.for("can.getValue");
 var onValueSymbol = canSymbol.for("can.onValue");
 var setValueSymbol = canSymbol.for("can.setValue");
 
@@ -112,49 +117,43 @@ function Bind(options) {
 	// either the child or parent is updated, we increase their respective
 	// semaphore so that if it’s two-way binding, then the “other” observable
 	// will only update if the total count for both semaphores is less than or
-	// equal to twice the number of allowed cycles.
-	this.childSemaphore = new Semaphore();
-	this.parentSemaphore = new Semaphore();
+	// equal to twice the number of cycles (because a cycle means two updates).
+	var childSemaphore = new Semaphore();
+	var parentSemaphore = new Semaphore();
 
-	// options.sticky === "childSticksToParent": after the child sets the parent,
-	// check to see if the parent matches the child; if not, then set the child to
-	// the parent’s value. This is used in cases where the parent modifies its own
-	// value and the child should be kept in sync with it.
-	//!steal-remove-start
-	if (options.sticky === "childSticksToParent" && options.cycles > 0) {
-		throw new Error("childSticksToParent && cycles > 0 is not yet tested");
+	// Determine if this is a one-way or two-way binding; by default, accept
+	// whatever options are passed in, but if they’re not defined, then check for
+	// the getValue and setValue symbols on the child and parent values.
+	var childToParent = true;
+	if (typeof options.childToParent === "boolean") {
+		// Always let the option override any checks
+		childToParent = options.childToParent;
+	} else if (options.child[getValueSymbol] == null) {
+		// Child to parent won’t work if we can’t get the child’s value
+		childToParent = false;
+	} else if (options.setParent === undefined && options.parent[setValueSymbol] == null) {
+		// Child to parent won’t work if we can’t set the parent’s value
+		childToParent = false;
 	}
-	//!steal-remove-end
-
-	// options.sticky === "parentSticksToChild": after the parent sets the child,
-	// check to see if the child matches the parent; if not, then set the parent
-	// to the child’s value. This is used in cases where the child modifies its
-	// own value and the parent should be kept in sync with it.
-	//!steal-remove-start
-	if (options.sticky === "parentSticksToChild") {
-		throw new Error("parentSticksToChild is not yet tested");
+	var parentToChild = true;
+	if (typeof options.parentToChild === "boolean") {
+		// Always let the option override any checks
+		parentToChild = options.parentToChild;
+	} else if (options.parent[getValueSymbol] == null) {
+		// Parent to child won’t work if we can’t get the parent’s value
+		parentToChild = false;
+	} else if (options.setChild === undefined && options.child[setValueSymbol] == null) {
+		// Parent to child won’t work if we can’t set the child’s value
+		parentToChild = false;
 	}
-	//!steal-remove-end
-
-	// Determine if the child and/or parent should be updated
-	// Check if setParent was provided or the parent has setValue
-	this._childToParent = !!(options.setParent || options.parent[setValueSymbol]);
-	// If it does and the option is passed in, use the option’s value
-	if (this._childToParent === true && typeof options.childToParent === "boolean") {
-		this._childToParent = options.childToParent;
-	}
-	// Check if setChild was provided or the child has setValue
-	this._parentToChild = !!(options.setChild || options.child[setValueSymbol]);
-	// If it does and the option is passed in, use the option’s value
-	if (this._parentToChild === true && typeof options.parentToChild === "boolean") {
-		this._parentToChild = options.parentToChild;
-	}
-	if (this._childToParent === false && this._parentToChild === false) {
+	if (childToParent === false && parentToChild === false) {
 		throw new Error("Neither the child nor parent will be updated; this is a no-way binding");
 	}
+	this._childToParent = childToParent;
+	this._parentToChild = parentToChild;
 
-	// Custom child & parent setters can be supplied;
-	// if they aren’t provided, then create our own.
+	// Custom child & parent setters can be supplied; if they aren’t provided,
+	// then create our own.
 	if (options.setChild === undefined) {
 		options.setChild = defaultSetChild.bind(options);
 	}
@@ -165,8 +164,6 @@ function Bind(options) {
 	// Bind the methods to this instance
 	this.start = this.start.bind(this);
 	this.stop = this.stop.bind(this);
-	this.updateChild = this.updateChild.bind(this);
-	this.updateParent = this.updateParent.bind(this);
 
 	// Set the observables’ priority
 	if (options.priority !== undefined) {
@@ -174,12 +171,88 @@ function Bind(options) {
 		canReflect.setPriority(options.parent, options.priority);
 	}
 
+	// These variables keep track of how many updates are allowed in a cycle.
+	// cycles is multipled by two because one update is allowed for each side of
+	// the binding, child and parent. One more update is allowed depending on the
+	// sticky option; if it’s sticky, then one more update needs to be allowed.
+	var allowedUpdates = options.cycles * 2;
+	var allowedChildUpdates = allowedUpdates + (options.sticky === "childSticksToParent" ? 1 : 0);
+	var allowedParentUpdates = allowedUpdates + (options.sticky === "parentSticksToChild" ? 1 : 0);
+
 	// This keeps track of whether we’re bound to the child and/or parent; this
 	// allows startParent() to be called first and on() can be called later to
-	// finish setting up the child binding.
-	this._bindingState = {
+	// finish setting up the child binding. This is also checked when updating
+	// values; if stop() has been called but updateValue() is called, then we
+	// ignore the update.
+	var bindingState = this._bindingState = {
 		child: false,
 		parent: false
+	};
+
+	// This is the listener that’s called when the parent changes
+	this.updateChild = function(newValue) {
+		updateValue({
+			bindingState: bindingState,
+			newValue: newValue,
+
+			// Some options used for debugging
+			debugObservableName: "child",
+			debugPartnerName: "parent",
+
+			// Main observable values
+			observable: options.child,
+			setValue: options.setChild,
+			semaphore: childSemaphore,
+
+			// If the sum of the semaphores is less than or equal to this number, then
+			// it’s ok to update the child with the new value.
+			allowedUpdates: allowedChildUpdates,
+
+			// If options.sticky === "parentSticksToChild", then after the parent sets
+			// the child, check to see if the child matches the parent; if not, then
+			// set the parent to the child’s value. This is used in cases where the
+			// child modifies its own value and the parent should be kept in sync with
+			// the child.
+			sticky: options.sticky === "parentSticksToChild",
+
+			// Partner observable values
+			partner: options.parent,
+			setPartner: options.setParent,
+			partnerSemaphore: parentSemaphore
+		});
+	};
+
+	// This is the listener that’s called when the child changes
+	this.updateParent = function(newValue) {
+		updateValue({
+			bindingState: bindingState,
+			newValue: newValue,
+
+			// Some options used for debugging
+			debugObservableName: "parent",
+			debugPartnerName: "child",
+
+			// Main observable values
+			observable: options.parent,
+			setValue: options.setParent,
+			semaphore: parentSemaphore,
+
+			// If the sum of the semaphores is less than or equal to this number, then
+			// it’s ok to update the parent with the new value.
+			allowedUpdates: allowedParentUpdates,
+
+			// If options.sticky === "childSticksToParent", then after the child sets
+			// the parent, check to see if the parent matches the child; if not, then
+			// set the child to the parent’s value. This is used in cases where the
+			// parent modifies its own value and the child should be kept in sync with
+			// the parent.
+			sticky: options.sticky === "childSticksToParent",
+
+			// Partner observable values
+			partner: options.child,
+			setPartner: options.setChild,
+			partnerSemaphore: childSemaphore
+		});
 	};
 
 	//!steal-remove-start
@@ -244,7 +317,7 @@ Object.assign(Bind.prototype, {
 			childValue = canReflect.getValue(options.child);
 			this.updateParent(childValue);
 
-		} else {
+		} else if (this._parentToChild === true) {
 			// One-way parent -> child, so update the child
 			// Check if updating the child is allowed
 			if (options.onInitDoNotUpdateChild === false) {
@@ -274,87 +347,40 @@ Object.assign(Bind.prototype, {
 
 	// Turn off all the bindings
 	stop: function() {
+		var bindingState = this._bindingState;
 		var options = this._options;
 
 		// Turn off the parent listener
-		if (this._bindingState.parent === true && this._parentToChild === true) {
-			this._bindingState.parent = false;
+		if (bindingState.parent === true && this._parentToChild === true) {
+			bindingState.parent = false;
 			turnOffListeningAndUpdate(options.parent, options.child, this.updateChild, options.queue);
 		}
 
 		// Turn off the child listener
-		if (this._bindingState.child === true && this._childToParent === true) {
-			this._bindingState.child = false;
+		if (bindingState.child === true && this._childToParent === true) {
+			bindingState.child = false;
 			turnOffListeningAndUpdate(options.child, options.parent, this.updateParent, options.queue);
 		}
-	},
-
-	// This is the listener that’s called when the parent changes
-	updateChild: function(newValue) {
-		var options = this._options;
-
-		updateValue({
-			newValue: newValue,
-
-			// Main observable values
-			observable: options.child,
-			setValue: options.setChild,
-			semaphore: this.childSemaphore,
-
-			// If the sum of the semaphores is less than or equal to this number, then
-			// it’s ok to update the child with the new value.
-			allowedUpdates: options.cycles * 2 + (options.sticky === "childSticksToParent" ? 1 : 0),
-
-			// If sticky is true, then after setting the child, check to see if its
-			// value is different from what it was set to; if it is, then set the
-			// parent to match the child’s new(est) value.
-			sticky: options.sticky === "parentSticksToChild",
-
-			// Partner observable values
-			partner: options.parent,
-			setPartner: options.setParent,
-			partnerSemaphore: this.parentSemaphore
-		});
-	},
-
-	// This is the listener that’s called when the child changes
-	updateParent: function(newValue) {
-		var options = this._options;
-
-		updateValue({
-			newValue: newValue,
-
-			// Main observable values
-			observable: options.parent,
-			setValue: options.setParent,
-			semaphore: this.parentSemaphore,
-
-			// If the sum of the semaphores is less than or equal to this number, then
-			// it’s ok to update the parent with the new value.
-			allowedUpdates: options.cycles * 2 + (options.sticky === "parentSticksToChild" ? 1 : 0),
-
-			// If sticky is true, then after setting the parent, check to see if its
-			// value is different from what it was set to; if it is, then set the
-			// child to match the parent’s new(est) value.
-			sticky: options.sticky === "childSticksToParent",
-
-			// Partner observable values
-			partner: options.child,
-			setPartner: options.setChild,
-			partnerSemaphore: this.childSemaphore
-		});
 	}
 
 });
 
 // updateValue is a helper function that’s used by updateChild and updateParent
 function updateValue(args) {
-	var semaphore = args.semaphore;
 
-	// Check the semaphore first; if this change is happening because the partner
+	// Check to see whether the binding is active; ignore updates if it isn’t active
+	var bindingState = args.bindingState;
+	if (bindingState.child === false && bindingState.parent === false) {
+		// We don’t warn the user about this because it’s a common occurrence in
+		// can-stache-bindings, e.g. {{#if value}}<input value:bind="value"/>{{/if}}
+		return;
+	}
+
+	// Now check the semaphore; if this change is happening because the partner
 	// observable was just updated, we only want to update this observable again
 	// if the total count for both semaphores is less than or equal to the number
 	// of allowed updates.
+	var semaphore = args.semaphore;
 	if ((semaphore.value + args.partnerSemaphore.value) <= args.allowedUpdates) {
 		queues.batch.start();
 
@@ -368,13 +394,13 @@ function updateValue(args) {
 		semaphore.increment();
 
 		// Decrease the semaphore after all other updates have occurred
-		queues.mutateQueue.enqueue(semaphore.decrement, semaphore, [], {});
+		queues.mutateQueue.enqueue(semaphore.decrement, semaphore, []);
 
 		queues.batch.stop();
 
 		// Stickiness is used in cases where the call to args.setValue above might
 		// have resulted in the observable being set to a different value than what
-		// was passed into this function (args.newValue). If sticy:true, then set
+		// was passed into this function (args.newValue). If sticky:true, then set
 		// the partner observable’s value so they’re kept in sync.
 		if (args.sticky) {
 			var observableValue = canReflect.getValue(args.observable);
@@ -382,6 +408,27 @@ function updateValue(args) {
 				args.setPartner(observableValue);
 			}
 		}
+
+	} else {
+		// It’s natural for this “else” block to be hit in two-way bindings; as an
+		// example, if a parent gets set and the child gets updated, the child’s
+		// listener to update the parent will be called, but it’ll be ignored if we
+		// don’t want cycles. HOWEVER, if this gets called and the parent is not the
+		// same value as the child, then their values are going to be out of sync,
+		// probably unintentionally. This is worth pointing out to developers
+		// because it can cause unexpected behavior… some people call those bugs. :)
+
+		//!steal-remove-start
+		var currentValue = canReflect.getValue(args.observable);
+		if (currentValue !== args.newValue) {
+			var warningParts = [
+				"can-bind updateValue: attempting to update " + args.debugObservableName + " " + canReflect.getName(args.observable) + " to new value: %o",
+				"…but the " + args.debugObservableName + " semaphore is at " + semaphore.value + " and the " + args.debugPartnerName + " semaphore is at " + args.partnerSemaphore.value + ". The number of allowed updates is " + args.allowedUpdates + ".",
+				"The " + args.debugObservableName + " value will remain unchanged; it’s currently: %o"
+			];
+			canLog.warn(warningParts.join("\n"), args.newValue, currentValue);
+		}
+		//!steal-remove-end
 	}
 }
