+++ package.json
@@ -14,7 +14,7 @@
   },
   "scripts": {
     "build": "node build.js",
-    "jshint": "jshint ./*.js --config",
+    "jshint": "jshint ./*.js ./test/*.js --config",
     "test": "npm run jshint && npm run testee",
     "testee": "testee test/test.html --browsers firefox",
     "preversion": "npm test && npm run build",
@@ -45,6 +45,8 @@
     "can-observation": "^4.0.1",
     "can-simple-map": "^4.0.1",
     "can-simple-observable": "^2.0.4",
+    "can-test-helpers": "^1.1.2",
+    "can-value": "<2.0.0",
     "jshint": "^2.9.1",
     "steal": "^1.6.5",
     "steal-qunit": "^1.0.1",
